feat: add typing indicator component animation and extract to file
